fix: remove shared dir, deployment does not work
