+++ apps/api/package.json
@@ -14,13 +14,12 @@
     "start": "NODE_ENV=production npx -y prisma migrate deploy && npx prisma generate && npx prisma db seed && node index.js"
   },
   "dependencies": {
-    "shared": "workspace:*",
     "@breejs/ts-worker": "2.0.0",
-    "@fastify/autoload": "5.0.0",
-    "@fastify/cookie": "7.0.0",
+    "@fastify/autoload": "5.1.0",
+    "@fastify/cookie": "7.1.0",
     "@fastify/cors": "8.0.0",
     "@fastify/env": "4.0.0",
-    "@fastify/jwt": "6.2.0",
+    "@fastify/jwt": "6.3.1",
     "@fastify/static": "6.4.0",
     "@iarna/toml": "2.2.5",
     "@prisma/client": "3.15.2",
@@ -33,8 +32,8 @@
     "dayjs": "1.11.3",
     "dockerode": "3.3.2",
     "dotenv-extended": "2.9.0",
-    "fastify": "4.2.0",
-    "fastify-plugin": "3.0.1",
+    "fastify": "4.2.1",
+    "fastify-plugin": "4.0.0",
     "generate-password": "1.7.0",
     "get-port": "6.1.2",
     "got": "12.1.0",
@@ -48,10 +47,10 @@
     "unique-names-generator": "4.7.1"
   },
   "devDependencies": {
-    "@types/node": "18.0.3",
+    "@types/node": "18.0.4",
     "@types/node-os-utils": "1.3.0",
-    "@typescript-eslint/eslint-plugin": "5.30.5",
-    "@typescript-eslint/parser": "5.30.5",
+    "@typescript-eslint/eslint-plugin": "5.30.6",
+    "@typescript-eslint/parser": "5.30.6",
     "esbuild": "0.14.49",
     "eslint": "8.19.0",
     "eslint-config-prettier": "8.5.0",
