+++ apps/api/src/routes/webhooks/traefik/handlers.ts
@@ -1,6 +1,5 @@
 import { FastifyRequest } from "fastify";
-import { supportedServiceTypesAndVersions } from "shared";
-import { asyncExecShell, errorHandler, getDomain, isDev, listServicesWithIncludes, prisma } from "../../../lib/common";
+import { asyncExecShell, errorHandler, getDomain, isDev, listServicesWithIncludes, prisma, supportedServiceTypesAndVersions } from "../../../lib/common";
 import { getEngine } from "../../../lib/docker";
 import { TraefikOtherConfiguration } from "./types";
