--- apps/shared/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "shared",
-  "version": "1.0.0",
-  "main": "build/index.js",
-  "scripts": {
-    "dev": "nodemon",
-    "build": "rimraf build && esbuild index.ts --minify=true --platform=node --outdir=build --format=cjs"
-  },
-  "devDependencies": {
-    "esbuild": "0.14.49",
-    "nodemon": "2.0.19",
-    "rimraf": "3.0.2"
-  }
-}
