+++ apps/ui/package.json
@@ -15,12 +15,11 @@
 		"format": "prettier --write --plugin-search-dir=. ."
 	},
 	"devDependencies": {
-		"@playwright/test": "1.23.1",
-		"@sveltejs/adapter-auto": "1.0.0-next.53",
-		"@sveltejs/kit": "1.0.0-next.359",
+		"@playwright/test": "1.23.3",
+		"@sveltejs/kit": "1.0.0-next.375",
 		"@types/js-cookie": "3.0.2",
-		"@typescript-eslint/eslint-plugin": "5.30.5",
-		"@typescript-eslint/parser": "5.30.5",
+		"@typescript-eslint/eslint-plugin": "5.30.6",
+		"@typescript-eslint/parser": "5.30.6",
 		"autoprefixer": "10.4.7",
 		"eslint": "8.19.0",
 		"eslint-config-prettier": "8.5.0",
@@ -28,19 +27,18 @@
 		"postcss": "8.4.14",
 		"prettier": "2.7.1",
 		"prettier-plugin-svelte": "2.7.0",
-		"svelte": "3.48.0",
+		"svelte": "3.49.0",
 		"svelte-check": "2.8.0",
 		"svelte-preprocess": "4.10.7",
-		"tailwindcss": "3.1.4",
+		"tailwindcss": "3.1.6",
 		"tailwindcss-scrollbar": "0.1.0",
 		"tslib": "2.4.0",
 		"typescript": "4.7.4",
-		"vite": "^2.9.13"
+		"vite": "^3.0.0"
 	},
 	"type": "module",
 	"dependencies": {
-		"shared": "workspace:*",
-		"@sveltejs/adapter-static": "1.0.0-next.34",
+		"@sveltejs/adapter-static": "1.0.0-next.36",
 		"@zerodevx/svelte-toast": "0.7.2",
 		"cuid": "2.1.8",
 		"js-cookie": "3.0.1",
