+++ apps/ui/src/routes/services/[id]/configuration/version.svelte
@@ -31,8 +31,7 @@
 	import { goto } from '$app/navigation';
 	import { get, post } from '$lib/api';
 	import { t } from '$lib/translations';
-	import { errorNotification } from '$lib/common';
-	import { supportedServiceTypesAndVersions } from 'shared/index';
+	import { errorNotification, supportedServiceTypesAndVersions } from '$lib/common';
 
 	const { id } = $page.params;
 	const from = $page.url.searchParams.get('from');
