+++ apps/ui/vite.config.js
@@ -7,6 +7,7 @@ export default {
         'GITPOD_WORKSPACE_URL': JSON.stringify(process.env.GITPOD_WORKSPACE_URL)
     },
     server: {
+        port: 3000,
         hmr: process.env.GITPOD_WORKSPACE_URL
         ? {
             // Due to port fowarding, we have to replace
